+++ geneticengine/algorithms/callbacks/csv_callback.py
@@ -28,7 +28,6 @@ class CSVCallback(Callback):
         self.filter_population = filter_population
         self.time = 0.0
         self.test_data = test_data
-        self.save_productions = save_productions
         self.only_record_best_ind = only_record_best_ind
         self.extra_columns = extra_columns or {}
         self.write_header()
@@ -47,10 +46,6 @@ class CSVCallback(Callback):
             "Execution Time",
             "Seed",
         ]
-        if self.test_data:
-            row.append("Test fitness")
-        if self.save_productions:
-            row.append("productions")
         for name, _ in self.extra_columns:
             row.append(name)
         self.writer.writerow(row)
@@ -78,10 +73,6 @@ class CSVCallback(Callback):
                 self.time,
                 gp.random_source.seed,
             ]
-            if self.test_data:
-                row.append(self.test_data(phenotype))
-            if self.save_productions:
-                row.append(ind.count_prods(gp.representation.genotype_to_phenotype, gp.representation.grammar))
 
             for (name, fun) in self.extra_columns.items():
                 row.append(fun(generation, population, time, gp, ind))
